+++ src/search/hamming/kernel.h
@@ -37,15 +37,15 @@ using ::DISPATCH_ARCH::FingerPrint;
 
 namespace Search { namespace DISPATCH_ARCH {
 
-static void all_vs_all(const array<char, 48>* __restrict a, size_t na, const array<char, 48>* __restrict b, size_t nb, HitField& out, const unsigned hamming_filter_id) {
-	const size_t na2 = na & ~size_t(3);
-	size_t i = 0;
+static void all_vs_all(const array<char, 48>* __restrict a, uint_fast32_t na, const array<char, 48>* __restrict b, uint_fast32_t nb, HitField& out, const unsigned hamming_filter_id) {
+	const uint_fast32_t na2 = na & ~uint_fast32_t(3);
+	uint_fast32_t i = 0;
 	for (; i < na2; i += 4) {
 		const FingerPrint e1(a[i]);
 		const FingerPrint e2(a[i + 1]);
 		const FingerPrint e3(a[i + 2]);
 		const FingerPrint e4(a[i + 3]);
-		for (size_t j = 0; j < nb; ++j) {
+		for (uint_fast32_t j = 0; j < nb; ++j) {
 			const FingerPrint fb(b[j]);
 			out.set(i, j, e1.match(fb) >= hamming_filter_id);
 			out.set(i+1, j, e2.match(fb) >= hamming_filter_id);
@@ -55,13 +55,13 @@ static void all_vs_all(const array<char, 48>* __restrict a, size_t na, const arr
 	}
 	for (; i < na; ++i) {
 		const FingerPrint e(a[i]);
-		for (size_t j = 0; j < nb; ++j)
+		for (uint_fast32_t j = 0; j < nb; ++j)
 			out.set(i, j, e.match(FingerPrint(b[j])) >= hamming_filter_id);
 	}
 }
 
 template<typename SeedLoc>
-static void FLATTEN stage1(const SeedLoc* __restrict q, int32_t nq, const SeedLoc* __restrict s, int32_t ns, WorkSet& work_set)
+static void FLATTEN stage1(const SeedLoc* __restrict q, uint_fast32_t nq, const SeedLoc* __restrict s, uint_fast32_t ns, WorkSet& work_set)
 {
 #ifdef __APPLE__
 	thread_local Container vq, vs;
@@ -69,15 +69,15 @@ static void FLATTEN stage1(const SeedLoc* __restrict q, int32_t nq, const SeedLo
 	Container& vq = work_set.vq, & vs = work_set.vs;
 #endif
 
-	const int32_t tile_size = config.tile_size;
+	const uint_fast32_t tile_size = config.tile_size;
 	::DISPATCH_ARCH::load_fps(s, ns, vs, work_set.cfg.target->seqs());
 	work_set.stats.inc(Statistics::SEED_HITS, nq * ns);
 	::DISPATCH_ARCH::load_fps(q, nq, vq, work_set.cfg.query->seqs());
-	const int32_t qs = (int32_t)vq.size(), ss = (int32_t)vs.size();
-	for (int32_t i = 0; i < qs; i += tile_size) {
-		for (int32_t j = 0; j < ss; j += tile_size) {
-			const size_t tq = std::min(tile_size, qs - i);
-			const size_t ts = std::min(tile_size, ss - j);
+	const uint_fast32_t qs = (uint_fast32_t)vq.size(), ss = (uint_fast32_t)vs.size();
+	for (uint_fast32_t i = 0; i < qs; i += tile_size) {
+		for (uint_fast32_t j = 0; j < ss; j += tile_size) {
+			const uint_fast32_t tq = std::min(tile_size, qs - i);
+			const uint_fast32_t ts = std::min(tile_size, ss - j);
 			work_set.hits.init(tq, ts);
 			all_vs_all(vq.data() + i, tq, vs.data() + j, ts, work_set.hits, work_set.cfg.hamming_filter_id);
 			search_tile(work_set.hits, i, j, q, s, work_set);
