+++ src/search/hamming/kernel_self.h
@@ -37,9 +37,9 @@ using std::vector;
 
 namespace Search { namespace DISPATCH_ARCH {
 	
-static void all_vs_all_self(const array<char, 48>* __restrict a, uint32_t na, HitField& out, unsigned hamming_filter_id) {
-	const size_t na2 = na & ~size_t(3);
-	size_t i = 0;
+static void all_vs_all_self(const array<char, 48>* __restrict a, uint_fast32_t na, HitField& out, unsigned hamming_filter_id) {
+	const uint_fast32_t na2 = na & ~uint_fast32_t(3);
+	uint_fast32_t i = 0;
 	for (; i < na2; i += 4) {
 		const FingerPrint e1(a[i]);
 		const FingerPrint e2(a[i + 1]);
@@ -51,7 +51,7 @@ static void all_vs_all_self(const array<char, 48>* __restrict a, uint32_t na, Hi
 		out.set(i + 1, i + 2, e2.match(e3) >= hamming_filter_id);
 		out.set(i + 1, i + 3, e2.match(e4) >= hamming_filter_id);
 		out.set(i + 2, i + 3, e3.match(e4) >= hamming_filter_id);
-		for (uint32_t j = i + 4; j < na; ++j) {
+		for (uint_fast32_t j = i + 4; j < na; ++j) {
 			const FingerPrint fa(a[j]);
 			out.set(i, j, e1.match(fa) >= hamming_filter_id);
 			out.set(i+1, j, e2.match(fa) >= hamming_filter_id);
@@ -61,13 +61,13 @@ static void all_vs_all_self(const array<char, 48>* __restrict a, uint32_t na, Hi
 	}
 	for (; i < na; ++i) {
 		const FingerPrint e(a[i]);
-		for (uint32_t j = i + 1; j < na; ++j)
+		for (uint_fast32_t j = i + 1; j < na; ++j)
 			out.set(i, j, e.match(FingerPrint(a[j])) >= hamming_filter_id);
 	}
 }
 
 template<typename SeedLoc>
-static void FLATTEN stage1_self(const SeedLoc* q, int32_t nq, const SeedLoc* s, int32_t ns, WorkSet& work_set)
+static void FLATTEN stage1_self(const SeedLoc* q, uint_fast32_t nq, const SeedLoc* s, uint_fast32_t ns, WorkSet& work_set)
 {
 #ifdef __APPLE__
 	thread_local Container vs;
@@ -75,19 +75,19 @@ static void FLATTEN stage1_self(const SeedLoc* q, int32_t nq, const SeedLoc* s, 
 	Container& vs = work_set.vs;
 #endif
 
-	const int32_t tile_size = config.tile_size;
+	const uint_fast32_t tile_size = config.tile_size;
 	::DISPATCH_ARCH::load_fps(s, ns, vs, work_set.cfg.target->seqs());
 
 	work_set.stats.inc(Statistics::SEED_HITS, ns * (ns - 1) / 2);
-	const int32_t ss = (int32_t)vs.size();
-	for (int32_t i = 0; i < ss; i += tile_size) {
-		const size_t tq = std::min(tile_size, ss - i);
+	const uint_fast32_t ss = (uint_fast32_t)vs.size();
+	for (uint_fast32_t i = 0; i < ss; i += tile_size) {
+		const uint_fast32_t tq = std::min(tile_size, ss - i);
 		work_set.hits.init(tq, tq);
 		all_vs_all_self(vs.data() + i, tq, work_set.hits, work_set.cfg.hamming_filter_id);
 		search_tile(work_set.hits, i, i, s, s, work_set);
-		for (int32_t j = i + tile_size; j < ss; j += tile_size) {
-			const size_t tq = std::min(tile_size, ss - i);
-			const size_t ts = std::min(tile_size, ss - j);
+		for (uint_fast32_t j = i + tile_size; j < ss; j += tile_size) {
+			const uint_fast32_t tq = std::min(tile_size, ss - i);
+			const uint_fast32_t ts = std::min(tile_size, ss - j);
 			work_set.hits.init(tq, ts);
 			all_vs_all(vs.data() + i, tq, vs.data() + j, ts, work_set.hits, work_set.cfg.hamming_filter_id);
 			search_tile(work_set.hits, i, j, s, s, work_set);
